requirements.txt
# --- Framework principal ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- ORM y Base de datos ---
sqlalchemy==2.0.32
mysql-connector-python==9.0.0

# --- WebSockets y tiempo real ---
python-socketio[asgi]==5.11.4
python-socketio[asyncio_client]==5.11.4
python-engineio==4.10.1

# --- Subida de archivos y formularios ---
python-multipart==0.0.9

# --- Seguridad y autenticación ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- Variables de entorno (.env) ---
python-dotenv==1.0.1

# --- Validación de datos ---
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# --- Logging y utilidades ---
loguru==0.7.2

# --- Dependencias útiles solo en desarrollo ---
watchfiles==0.23.0
pytest==8.3.3
